requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
click==8.2.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
filelock==3.19.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.3.8
psycopg2-binary==2.9.10
setuptools==80.9.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
uvicorn==0.35.0
virtualenv==20.34.0
whitenoise==6.9.0
